Add a `UnitVec3`, guaranteed to be of unit length. Used in e.g. `Hit`.
I'm not satisfied with the ergonomics of this newtype, and I currently
think that's at least partially Rust's fault. The struct is supposed to
be a very thin wrapper around `Vec3`s, and should ideally be usable
transparently wherever a `Vec3` is used in immutable contexts. This,
however, is not trivial without a lot of boilerplate. I might try to
improve the ergonomics in the future, but this suffices for now.